Updated code to remove code duplication
