+++ sources/packages/backend/apps/api/src/app.aest.module.ts
@@ -64,6 +64,7 @@ import {
   ConfirmationOfEnrollmentControllerService,
   OverawardAESTController,
   OverawardControllerService,
+  RestrictionControllerService,
 } from "./route-controllers";
 import { AuthModule } from "./auth/auth.module";
 import {
@@ -139,6 +140,7 @@ import {
     SFASPartTimeApplicationsService,
     StudentRestrictionService,
     RestrictionService,
+    RestrictionControllerService,
     ApplicationExceptionService,
     StudentScholasticStandingsService,
     ScholasticStandingControllerService,
