+++ sources/packages/backend/apps/api/src/route-controllers/index.ts
@@ -59,3 +59,4 @@ export * from "./overaward/overaward.institutions.controller";
 export * from "./overaward/overaward.students.controller";
 export * from "./student/student.institutions.controller";
 export * from "./restriction/restriction.institution.controller";
+export * from "./restriction/restriction.controller.service";
